feat(auth): move PasskeyPromptDialog to different place
