+++ lib/app/features/auth/views/components/auth_footer/auth_footer.dart
@@ -3,11 +3,11 @@
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
-import 'package:ion/app/components/auth/secured_by.dart';
 import 'package:ion/app/constants/links.dart';
 import 'package:ion/app/extensions/build_context.dart';
 import 'package:ion/app/extensions/num.dart';
 import 'package:ion/app/extensions/theme_data.dart';
+import 'package:ion/app/features/components/passkeys/secured_by.dart';
 import 'package:ion/app/services/browser/browser.dart';
 import 'package:ion/l10n/i10n.dart';
