+++ lib/app/features/components/passkeys/passkey_prompt_dialog.dart
@@ -2,14 +2,14 @@
 
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:ion/app/components/auth/secured_by.dart';
 import 'package:ion/app/components/progress_bar/ion_loading_indicator.dart';
 import 'package:ion/app/components/screen_offset/screen_bottom_offset.dart';
 import 'package:ion/app/extensions/extensions.dart';
+import 'package:ion/app/features/components/passkeys/secured_by.dart';
 import 'package:ion/app/features/core/providers/theme_mode_provider.dart';
 import 'package:ion/generated/assets.gen.dart';
 
-class PasskeyPromptDialog extends HookConsumerWidget {
+class PasskeyPromptDialog extends ConsumerWidget {
   const PasskeyPromptDialog({
     super.key,
   });
